backend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-multipart
openpyxl
python-dotenv
openai
requests
